entrance: widen EntranceInfo.scene to s16
Vanilla has 0x6E scenes, so one byte was ample. With custom scenes appended
the table has reached 0x7F, and the next id - 0x80 - read back as -128.

The failure is silent and indirect, not an error. A negative id fails every
`== SCENE_*` test in Play_Init, including the GridToolSceneRegistry check
that forces sceneLayer to 0 for grid-tool scenes (which use one entrance row
instead of vanilla's four). sceneLayer therefore keeps its day/night value
and Play_SpawnScene is handed gEntranceTable[entranceIndex + sceneLayer],
read past the end of the table. The entrance simply lands in a different
scene.

Nothing serialises EntranceInfo - saves store the entrance index, and
gEntranceTable is built at compile time from tables/entrance_table.h - so
there is no format consequence. Two randomizer locals that had copied the
field into an int8_t/s8 follow, and Grotto_GetRenamedGrottoIndexFromOriginal's
scene parameter with them; every grotto in grottoLoadTable is a vanilla
scene, so its comparison is unaffected.

Found while proving sturdy-bassoon#27, which needed a new scene to exist.
Note that nothing warns when an emitter registers a scene past the limit -
that gap wants its own issue. EntranceInfo.spawn is the same s8 and is 0 for
every custom scene today, but it is the same trap.

+++ soh/soh/Enhancements/randomizer/randomizer_check_tracker.cpp
@@ -2441,7 +2441,7 @@ void InternalRecalculateAvailableChecks(RandomizerRegion startingRegion, RandoAg
     if (startingRegion == RR_ROOT && entranceIndex >= 0 && entranceIndex < ENTR_MAX) {
         // Try to find a mapped entrance
         // e.g. ENTR_DEKU_TREE_0_1 (index 1) is not mapped, but ENTR_DEKU_TREE_ENTRANCE (index 0) is mapped
-        const int8_t scene = gEntranceTable[entranceIndex].scene;
+        const int16_t scene = gEntranceTable[entranceIndex].scene;
         for (; entranceIndex >= 0 && gEntranceTable[entranceIndex].scene == scene; entranceIndex--) {
             const auto entrance = Rando::EntranceShuffler::GetEntranceByIndex(entranceIndex);
             if (entrance != nullptr) {
